LayoutTests/platform/ios-simulator/webaudio/realtimeanalyser-fft-scaling-expected.txt

Test scaling of FFT data for AnalyserNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Actual FFT peak in the expected position (1)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 32
PASS Actual FFT peak in the expected position (2)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 64
PASS Actual FFT peak in the expected position (4)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 128
PASS Actual FFT peak in the expected position (8)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 256
PASS Actual FFT peak in the expected position (16)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 512
PASS Actual FFT peak in the expected position (32)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 1024
PASS Actual FFT peak in the expected position (64)
PASS Peak value is near 0 dBFS as expected
PASS Analyser correctly scaled FFT data of size 2048
PASS All Analyser tests passed.
PASS successfullyParsed is true

TEST COMPLETE
